use base colors as references
