+++ frontend/src/metabase/lib/colors/constants/dark.ts
@@ -1,7 +1,10 @@
 /* eslint-disable no-color-literals */
 import type { MetabaseThemeV2 } from "../types";
 
+import { getBaseColorsForThemeDefinitionOnly } from "./base-colors";
 import { DEFAULT_CHART_COLORS } from "./chart-colors";
+
+const baseColors = getBaseColorsForThemeDefinitionOnly();
 
 /**
  * The default dark theme for Metabase.
@@ -10,83 +13,82 @@ export const METABASE_DARK_THEME: MetabaseThemeV2 = {
   version: 2,
   chartColors: DEFAULT_CHART_COLORS,
   colors: {
-    "accent-gray-dark": "hsla(205, 63%, 5%, 1)",
-    "accent-gray-light": "hsla(205, 19%, 23%, 1)",
-    "accent-gray": "hsla(205, 19%, 23%, 1)",
-    "admin-navbar": "hsla(240, 43%, 33%, 1)",
-    "admin-navbar-secondary": "hsla(240, 46%, 58%, 1)",
-    "admin-navbar-inverse": "hsla(240, 46%, 58%, 1)",
-    "background-brand": "color-mix(in srgb, var(--mb-color-brand), black 75%)",
-    "background-disabled": "hsla(0, 0%, 100%, 0.10)",
-    "background-disabled-inverse": "hsla(204, 66%, 8%, 0.05)",
-    "background-error-secondary": "hsla(1, 75%, 17%, 1)",
+    "accent-gray-dark": baseColors.orion[110],
+    "accent-gray-light": baseColors.orion[80],
+    "accent-gray": baseColors.orion[80],
+    "admin-navbar": baseColors.octopus[80],
+    "admin-navbar-secondary": baseColors.octopus[60],
+    "admin-navbar-inverse": baseColors.octopus[60],
+    "background-brand": baseColors.brand[90],
+    "background-disabled": baseColors.orionAlphaInverse[10],
+    "background-disabled-inverse": baseColors.orionAlpha[10],
+    "background-error-secondary": baseColors.lobster[90],
     "background-hover":
       "color-mix(in srgb, var(--mb-color-brand) 21%, transparent)",
     "background-hover-light":
       "color-mix(in srgb, var(--mb-color-brand) 7%, transparent)",
-    "background-selected": "var(--mb-color-brand)",
-    "background-primary": "hsla(204, 66%, 8%, 1)",
-    "background-secondary": "hsla(205, 63%, 5%, 1)",
-    "background-tertiary": "hsla(205, 19%, 23%, 1)",
-    "background-primary-inverse": "hsla(195, 6%, 87%, 1)",
-    "background-secondary-inverse": "hsla(203, 6%, 73%, 1)",
-    "background-tertiary-inverse": "hsla(204, 12%, 32%, 1)",
-    overlay: "hsla(204, 66%, 8%, 0.74)",
-    "bg-ocean-alpha-light": "hsla(208, 95%, 42%, 0.21)",
-    "background-error": "hsla(1, 75%, 17%, 1)",
-    "background-success": "hsla(92, 62%, 10%, 1)",
+    "background-selected": baseColors.brand[40],
+    "background-primary": baseColors.orion[100],
+    "background-secondary": baseColors.orion[110],
+    "background-tertiary": baseColors.orion[80],
+    "background-primary-inverse": baseColors.orion[20],
+    "background-secondary-inverse": baseColors.orion[30],
+    "background-tertiary-inverse": baseColors.orion[70],
+    overlay: baseColors.orionAlpha[70],
+    "bg-ocean-alpha-light": baseColors.oceanAlpha[20],
+    "background-error": baseColors.lobster[90],
+    "background-success": baseColors.palm[90],
     "brand-alpha-04":
       "color-mix(in srgb, var(--mb-color-brand) 4%, transparent)",
     "brand-alpha-88":
       "color-mix(in srgb, var(--mb-color-brand) 88%, transparent)",
-    "brand-dark": "color-mix(in srgb, var(--mb-color-brand), white 35%)",
-    "brand-darker": "color-mix(in srgb, var(--mb-color-brand), white 70%)",
-    "brand-light": "color-mix(in srgb, var(--mb-color-brand), black 62%)",
-    "brand-lighter": "color-mix(in srgb, var(--mb-color-brand), black 75%)",
-    brand: "hsla(208, 72%, 60%, 1.00)",
-    danger: "hsla(358, 71%, 62%, 1)",
-    error: "hsla(358, 71%, 62%, 1)",
-    filter: "hsla(240, 69%, 74%, 1)",
-    focus: "hsla(208, 80%, 31%, 1.00)",
+    "brand-dark": baseColors.brand[30],
+    "brand-darker": baseColors.brand[20],
+    "brand-light": baseColors.brand[80],
+    "brand-lighter": baseColors.brand[90],
+    brand: baseColors.blue[40],
+    danger: baseColors.lobster[50],
+    error: baseColors.lobster[50],
+    filter: baseColors.octopus[40],
+    focus: baseColors.blue[70],
     "icon-primary": "var(--mb-color-brand)",
-    "icon-disabled": "hsla(0, 0%, 100%, 0.33)",
-    "icon-secondary": "hsla(0, 0%, 100%, 0.53)",
-    "metabase-brand": "hsla(208, 72%, 60%, 1.00)",
-    "saturated-blue": "hsla(208, 72%, 60%, 1)",
-    "saturated-green": "hsla(89, 47%, 45%, 1)",
-    "saturated-purple": "hsla(240, 69%, 74%, 1)",
-    "saturated-red": "hsla(1, 84%, 69%, 1)",
-    "saturated-yellow": "hsla(46, 81%, 52%, 1)",
-    shadow: "color-mix(in srgb, hsla(205, 63%, 5%, 1) 20%, transparent)",
-    "success-secondary": "hsla(89, 47%, 45%, 1)",
-    success: "hsla(89, 48%, 40%, 1)",
-    summarize: "hsla(89, 47%, 45%, 1)",
-    "switch-off": "hsla(0, 0%, 100%, 0.21)",
-    "syntax-parameters-active": "hsla(26, 79%, 13%, 1)",
-    "syntax-parameters": "hsla(26, 79%, 54%, 1)",
-    "text-brand": "var(--mb-color-brand)",
-    "text-tertiary": "hsla(0, 0%, 100%, 0.46)",
-    "text-tertiary-inverse": "hsla(204, 66%, 8%, 0.44)",
-    "text-hover": "color-mix(in srgb, var(--mb-color-brand), white 35%)",
-    "text-secondary-opaque": "hsla(195, 6%, 87%, 1)",
-    "text-primary": "hsla(0, 0%, 100%, 0.95)",
-    "text-primary-inverse": "hsla(204, 66%, 8%, 0.84)",
-    "text-secondary": "hsla(0, 0%, 100%, 0.69)",
-    "text-secondary-inverse": "hsla(204, 66%, 8%, 0.62)",
-    "text-selected": "hsla(0, 0%, 100%, 1.00)",
-    "tooltip-background-focused":
-      "color-mix(in srgb, hsla(204, 12%, 32%, 1) 50%, #000)",
-    "tooltip-background": "hsla(204, 12%, 32%, 1)",
-    "tooltip-text-secondary": "hsla(0, 0%, 100%, 0.69)",
-    "tooltip-text": "hsla(0, 0%, 100%, 0.95)",
-    warning: "hsla(46, 81%, 52%, 1)",
-    "background-warning": "hsla(46, 88%, 10%, 1)",
-    info: "hsla(203, 5%, 53%, 1)",
-    "background-info": "hsla(204, 34%, 14%, 1)",
-    white: "hsla(205, 63%, 5%, 1)",
-    border: "hsla(0, 0%, 100%, 0.21)",
-    "border-strong": "hsla(0, 0%, 100%, 0.53)",
-    "border-subtle": "hsla(0, 0%, 100%, 0.10)",
+    "icon-disabled": baseColors.orionAlphaInverse[30],
+    "icon-secondary": baseColors.orionAlphaInverse[50],
+    "metabase-brand": baseColors.blue[40],
+    "saturated-blue": baseColors.ocean[40],
+    "saturated-green": baseColors.palm[40],
+    "saturated-purple": baseColors.octopus[40],
+    "saturated-red": baseColors.lobster[40],
+    "saturated-yellow": baseColors.dubloon[30],
+    shadow: `color-mix(in srgb, ${baseColors.orion[110]} 20%, transparent)`,
+    "success-secondary": baseColors.palm[40],
+    success: baseColors.palm[50],
+    summarize: baseColors.palm[40],
+    "switch-off": baseColors.orionAlphaInverse[20],
+    "syntax-parameters-active": baseColors.mango[90],
+    "syntax-parameters": baseColors.mango[40],
+    "text-brand": baseColors.brand[40],
+    "text-tertiary": baseColors.orionAlphaInverse[40],
+    "text-tertiary-inverse": baseColors.orionAlpha[40],
+    "text-hover": baseColors.brand[30],
+    "text-secondary-opaque": baseColors.orion[20],
+    "text-primary": baseColors.orionAlphaInverse[80],
+    "text-primary-inverse": baseColors.orionAlpha[80],
+    "text-secondary": baseColors.orionAlphaInverse[60],
+    "text-secondary-inverse": baseColors.orionAlpha[60],
+    "text-selected": baseColors.white,
+    "tooltip-background-focused": `color-mix(in srgb, ${baseColors.orion[70]} 50%, #000)`,
+    "tooltip-background": baseColors.orion[70],
+    "tooltip-text-secondary": baseColors.orionAlphaInverse[60],
+    "tooltip-text": baseColors.orionAlphaInverse[80],
+    warning: baseColors.dubloon[30],
+    "background-warning": baseColors.dubloon[90],
+    info: baseColors.orion[50],
+    "background-info": baseColors.orion[90],
+    white: baseColors.orion[110],
+    border: baseColors.orionAlphaInverse[20],
+    "border-strong": baseColors.orionAlphaInverse[50],
+    "border-subtle": baseColors.orionAlphaInverse[10],
     copper: "#B87333",
     bronze: "#CD7F32",
     silver: "#C0C0C0",
