+++ frontend/src/metabase/lib/colors/constants/light.ts
@@ -1,7 +1,10 @@
 /* eslint-disable no-color-literals */
 import type { MetabaseThemeV2 } from "../types";
 
+import { getBaseColorsForThemeDefinitionOnly } from "./base-colors";
 import { DEFAULT_CHART_COLORS } from "./chart-colors";
+
+const baseColors = getBaseColorsForThemeDefinitionOnly();
 
 /**
  * The default light theme for Metabase.
@@ -10,84 +13,82 @@ export const METABASE_LIGHT_THEME: MetabaseThemeV2 = {
   version: 2,
   chartColors: DEFAULT_CHART_COLORS,
   colors: {
-    "accent-gray-dark": "hsla(195, 6%, 87%, 1)",
-    "accent-gray-light": "hsla(240, 11%, 98%, 1)",
-    "accent-gray": "hsla(240, 4%, 95%, 1)",
-    "admin-navbar": "hsla(240, 46%, 58%, 1)",
-    "admin-navbar-secondary": "hsla(240, 69%, 74%, 1)",
-    "admin-navbar-inverse": "hsla(240, 43%, 33%, 1)",
-    "background-brand": "color-mix(in srgb, var(--mb-color-brand), white 91%)",
-    "background-disabled": "hsla(204, 66%, 8%, 0.05)",
-    "background-disabled-inverse": "hsla(0, 0%, 100%, 0.10)",
-    "background-error-secondary": "hsla(0, 100%, 99%, 1)",
+    "accent-gray-dark": baseColors.orion[20],
+    "accent-gray-light": baseColors.orion[5],
+    "accent-gray": baseColors.orion[10],
+    "admin-navbar": baseColors.octopus[60],
+    "admin-navbar-secondary": baseColors.octopus[40],
+    "admin-navbar-inverse": baseColors.octopus[80],
+    "background-brand": baseColors.brand[10],
+    "background-disabled": baseColors.orionAlpha[10],
+    "background-disabled-inverse": baseColors.orionAlphaInverse[10],
+    "background-error-secondary": baseColors.lobster[5],
     "background-hover":
       "color-mix(in srgb, var(--mb-color-brand) 21%, transparent)",
     "background-hover-light":
       "color-mix(in srgb, var(--mb-color-brand) 7%, transparent)",
-    "background-selected":
-      "color-mix(in srgb, var(--mb-color-brand), black 12%)",
-    "background-primary": "hsla(0, 0%, 100%, 1.00)",
-    "background-secondary": "hsla(240, 11%, 98%, 1)",
-    "background-tertiary": "hsla(240, 4%, 95%, 1)",
-    "background-primary-inverse": "hsla(205, 19%, 23%, 1)",
-    "background-secondary-inverse": "hsla(204, 12%, 32%, 1)",
-    "background-tertiary-inverse": "hsla(205, 6%, 60%, 1)",
-    overlay: "hsla(204, 66%, 8%, 0.62)",
-    "bg-ocean-alpha-light": "hsla(208, 95%, 42%, 0.03)",
-    "background-error": "hsla(0, 76%, 97%, 1)",
-    "background-success": "hsla(93, 73%, 97%, 1)",
+    "background-selected": baseColors.brand[50],
+    "background-primary": baseColors.white,
+    "background-secondary": baseColors.orion[5],
+    "background-tertiary": baseColors.orion[10],
+    "background-primary-inverse": baseColors.orion[80],
+    "background-secondary-inverse": baseColors.orion[70],
+    "background-tertiary-inverse": baseColors.orion[40],
+    overlay: baseColors.orionAlpha[60],
+    "bg-ocean-alpha-light": baseColors.oceanAlpha[5],
+    "background-error": baseColors.lobster[10],
+    "background-success": baseColors.palm[5],
     "brand-alpha-04":
       "color-mix(in srgb, var(--mb-color-brand) 4%, transparent)",
     "brand-alpha-88":
       "color-mix(in srgb, var(--mb-color-brand) 88%, transparent)",
-    "brand-dark": "color-mix(in srgb, var(--mb-color-brand), black 28%)",
-    "brand-darker": "color-mix(in srgb, var(--mb-color-brand), black 47%)",
-    "brand-light": "color-mix(in srgb, var(--mb-color-brand), white 91%)",
-    "brand-lighter": "color-mix(in srgb, var(--mb-color-brand), white 96%)",
-    brand: "hsla(208, 72%, 60%, 1.00)",
-    danger: "hsla(358, 71%, 62%, 1)",
-    error: "hsla(358, 71%, 62%, 1)",
-    filter: "hsla(240, 65%, 69%, 1)",
-    focus: "hsla(209, 73%, 88%, 1.00)",
+    "brand-dark": baseColors.brand[60],
+    "brand-darker": baseColors.brand[70],
+    "brand-light": baseColors.brand[10],
+    "brand-lighter": baseColors.brand[5],
+    brand: baseColors.blue[40],
+    danger: baseColors.lobster[50],
+    error: baseColors.lobster[50],
+    filter: baseColors.octopus[50],
+    focus: baseColors.blue[20],
     "icon-primary": "var(--mb-color-brand)",
-    "icon-disabled": "hsla(204, 66%, 8%, 0.29)",
-    "icon-secondary": "hsla(204, 66%, 8%, 0.51)",
-    "metabase-brand": "hsla(208, 72%, 60%, 1.00)",
-    "saturated-blue": "hsla(208, 78%, 42%, 1)",
-    "saturated-green": "hsla(89, 48%, 32%, 1)",
-    "saturated-purple": "hsla(240, 46%, 58%, 1)",
-    "saturated-red": "hsla(1, 67%, 49%, 1)",
-    "saturated-yellow": "hsla(46, 81%, 52%, 1)",
-    shadow: "hsla(204, 66%, 8%, 0.17)",
-    "success-secondary": "hsla(89, 48%, 32%, 1)",
-    success: "hsla(89, 48%, 40%, 1)",
-    summarize: "hsla(89, 48%, 40%, 1)",
-    "switch-off": "hsla(204, 66%, 8%, 0.17)",
-    "syntax-parameters-active": "hsla(25, 100%, 95%, 1)",
-    "syntax-parameters": "hsla(26, 69%, 39%, 1)",
-    "text-brand": "color-mix(in srgb, var(--mb-color-brand), black 12%)",
-    "text-tertiary": "hsla(204, 66%, 8%, 0.44)",
-    "text-tertiary-inverse": "hsla(0, 0%, 100%, 0.46)",
-    "text-hover": "color-mix(in srgb, var(--mb-color-brand), black 28%)",
-    "text-secondary-opaque": "hsla(205, 8%, 43%, 1)",
-    "text-primary": "hsla(204, 66%, 8%, 0.84)",
-    "text-primary-inverse": "hsla(0, 0%, 100%, 0.95)",
-    "text-secondary": "hsla(204, 66%, 8%, 0.62)",
-    "text-secondary-inverse": "hsla(0, 0%, 100%, 0.69)",
-    "text-selected": "hsla(0, 0%, 100%, 1.00)",
-    "tooltip-background-focused":
-      "color-mix(in srgb, hsla(205, 19%, 23%, 1) 50%, #000)",
-    "tooltip-background": "hsla(205, 19%, 23%, 1)",
-    "tooltip-text-secondary": "hsla(0, 0%, 100%, 0.69)",
-    "tooltip-text": "hsla(0, 0%, 100%, 1.00)",
-    warning: "hsla(46, 81%, 52%, 1)",
-    "background-warning": "hsla(45, 100%, 96%, 1)",
-    info: "hsla(205, 6%, 60%, 1)",
-    "background-info": "hsla(240, 4%, 95%, 1)",
-    white: "hsla(0, 0%, 100%, 1.00)",
-    border: "hsla(195, 6%, 87%, 1)",
-    "border-strong": "hsla(204, 66%, 8%, 0.51)",
-    "border-subtle": "hsla(204, 66%, 8%, 0.05)",
+    "icon-disabled": baseColors.orionAlpha[30],
+    "icon-secondary": baseColors.orionAlpha[50],
+    "metabase-brand": baseColors.blue[40],
+    "saturated-blue": baseColors.ocean[60],
+    "saturated-green": baseColors.palm[60],
+    "saturated-purple": baseColors.octopus[60],
+    "saturated-red": baseColors.lobster[60],
+    "saturated-yellow": baseColors.dubloon[30],
+    shadow: baseColors.orionAlpha[20],
+    "success-secondary": baseColors.palm[60],
+    success: baseColors.palm[50],
+    summarize: baseColors.palm[50],
+    "switch-off": baseColors.orionAlpha[20],
+    "syntax-parameters-active": baseColors.mango[10],
+    "syntax-parameters": baseColors.mango[60],
+    "text-brand": baseColors.brand[50],
+    "text-tertiary": baseColors.orionAlpha[40],
+    "text-tertiary-inverse": baseColors.orionAlphaInverse[40],
+    "text-hover": baseColors.brand[60],
+    "text-secondary-opaque": baseColors.orion[60],
+    "text-primary": baseColors.orionAlpha[80],
+    "text-primary-inverse": baseColors.orionAlphaInverse[80],
+    "text-secondary": baseColors.orionAlpha[60],
+    "text-secondary-inverse": baseColors.orionAlphaInverse[60],
+    "text-selected": baseColors.white,
+    "tooltip-background-focused": `color-mix(in srgb, ${baseColors.orion[80]} 50%, #000)`,
+    "tooltip-background": baseColors.orion[80],
+    "tooltip-text-secondary": baseColors.orionAlphaInverse[60],
+    "tooltip-text": baseColors.white,
+    warning: baseColors.dubloon[30],
+    "background-warning": baseColors.dubloon[5],
+    info: baseColors.orion[40],
+    "background-info": baseColors.orion[10],
+    white: baseColors.white,
+    border: baseColors.orion[20],
+    "border-strong": baseColors.orionAlpha[50],
+    "border-subtle": baseColors.orionAlpha[10],
     copper: "#B87333",
     bronze: "#CD7F32",
     silver: "#C0C0C0",
